Fix relative URLs
When doing [text](#anchor), the link would always point to
https://reuse.software/#anchor instead of the anchor on the current
page. This was a bug in the theme.

Furthermore, I deleted the `site/public` directory, which is full of
compilation artifacts. I should have put that in a separate commit, but
my Git-fu is really lacking right now. I don't know how to fix it and
don't really want to put in the time to figure it out.