Simply comment header and .license explanation
Before this change, the explanation of comment headers and .license was
rather verbose because there was no concise way to say “plain text file
that can contain comments” or “not a plain text file or does not permit
the inclusion of comments”. This change adds terms to the glossary so
that the explanations of comment headers and .license files could be
shortened.

In the next commit, I’m going to add another reference to Commentable
Files. That’s the main motivation behind this change. I want to be able
to talk about Commentable Files without having to repeat the phrase
“plain text file that can contain comments”.

Signed-off-by: Carmen Bianca BAKKER <carmenbianca@fsfe.org>