Formally allow using .license files with Commentable Files
Dave Winer’s Rules for standards-makers says “If practice deviates from
the spec, change the spec” [1]. If a spec says to do one thing but
everyone does something else, then that rule in the spec is more of a
distraction then anything else.

Before this change, the REUSE Spec said “To implement this method, each
Commentable File MUST contain comments at the top of the file (comment
header) that declare that file's Copyright and Licensing Information.”
As far as I know, there’s no piece of software that actually enforces
that requirement.

For example, try running “reuse lint” on this repository:
<https://github.com/Jayman2000/reuse-compliance-test>. The REUSE Tool
will tell you that you’re compliant with the REUSE Spec, even though
there’s a .py.license file (The Python file itself could have just
contained a comment header).

This change aligns the REUSE Spec with actual practice by making this
requirement a SHOULD instead of a MUST. For the avoidance of doubt, this
change also explicitly says that “.license files MAY be used with
Commentable Files”.

The phrase “but it is still RECOMMENDED that comment headers be put
inside Commentable Files” was inspired by this comment:
<https://github.com/fsfe/reuse-docs/issues/122#issuecomment-1413513855>

Fixes #122.

[1]: <http://scripting.com/2017/05/09/rulesForStandardsmakers.html>

Signed-off-by: Carmen Bianca BAKKER <carmenbianca@fsfe.org>