Fix redirect to use English-language version
The problem: Sometimes, when Hugo generated the website, the `/spec/`
alias would point to the Ukrainian translation of the latest
specification. Exactly how or why this happens is unclear to me, but I
suspect it is because the Ukrainian Markdown file would also contain the
`aliases: ["/spec"]` frontmatter, and the Hugo compiler would somehow
toss a coin on which aliases frontmatter is authoritative.

Here, we get rid of the frontmatter method of doing aliases, and instead
use a 'layout' to effectively do the same thing.

Signed-off-by: Carmen Bianca BAKKER <carmenbianca@fsfe.org>